(feat): Modified all minutely jobs to use the cache_gate/redis solution.
